requirements.txt
Django==3.1.1
psycopg2-binary==2.8.6
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
drf-yasg==1.17.1
django-filter==2.3.0
gunicorn==20.0.4
sentry-sdk==0.17.3
django-ckeditor==6.0.0
drf-firebase-auth==0.0.8
dadata==20.7.0
